refactor: rename tickets dir
